Moved plotting script into plotting sub-directory
